Cleanup & PR-ification: Removed helper type class, cleaned up imports, moved into types module, added kind annotations to quantified tyvars, etc
